+++ src/abilities/items/cloakOfFlames.ts
@@ -1,5 +1,5 @@
 
-import { addScriptHook, W3TS_HOOK } from "@voces/w3ts";
+import { addScriptHook, W3TS_HOOK } from "w3ts";
 import { CLOAK_TYPE } from "shared";
 import { withTempGroup } from "util/temp";
 import { wrappedTriggerAddAction } from "util/emitLog";