+++ src/commands/index.ts
@@ -4,6 +4,7 @@ import "./control";
 import "./face";
 import "./gold";
 import "./lumber";
+import "./rotation";
 import "./sheepCommands";
 import "./uncontrol";
 import "./zoom";